+++ .github/workflows/e2e_api_tests.yml
@@ -427,26 +427,26 @@ jobs:
                       const script = require("./tests/pw/utils/gitTestSummary.ts")
                       return await script({github, context, core})
 
-            # Find PR comment
-            - name: Find PR comment by github-actions[bot]
-              id: find-comment
-              if: github.event_name == 'pull_request'
-              uses: peter-evans/find-comment@v3
-              with:
-                  issue-number: ${{ github.event.pull_request.number }}
-                  comment-author: 'github-actions[bot]'
-                  body-includes: Tests Summary
+            # # Find PR comment
+            # - name: Find PR comment by github-actions[bot]
+            #   id: find-comment
+            #   if: github.event_name == 'pull_request'
+            #   uses: peter-evans/find-comment@v3
+            #   with:
+            #       issue-number: ${{ github.event.pull_request.number }}
+            #       comment-author: 'github-actions[bot]'
+            #       body-includes: Tests Summary
 
-            # Post test summary as PR comment
-            - name: Create or update PR comment
-              if: github.event_name == 'pull_request'
-              uses: peter-evans/create-or-update-comment@v4
-              with:
-                  comment-id: ${{ steps.find-comment.outputs.comment-id }}
-                  issue-number: ${{ github.event.pull_request.number }}
-                  body: ${{ steps.prepare-test-summary.outputs.result }}
-                  reactions: hooray
-                  edit-mode: replace
+            # # Post test summary as PR comment
+            # - name: Create or update PR comment
+            #   if: github.event_name == 'pull_request'
+            #   uses: peter-evans/create-or-update-comment@v4
+            #   with:
+            #       comment-id: ${{ steps.find-comment.outputs.comment-id }}
+            #       issue-number: ${{ github.event.pull_request.number }}
+            #       body: ${{ steps.prepare-test-summary.outputs.result }}
+            #       reactions: hooray
+            #       edit-mode: replace
 
             # Upload final test artifact
             - name: Upload final test artifact
@@ -458,20 +458,3 @@ jobs:
                   if-no-files-found: ignore
                   retention-days: 7
 
-            # - name: Publish HTML report to GH Pages
-            #   if: always() && steps.generate-html-report.outcome == 'success'
-            #   uses: peaceiris/actions-gh-pages@v4
-            #   with:
-            #       github_token: ${{ secrets.GITHUB_TOKEN }}
-            #       publish_branch: gh-pages
-            #       publish_dir: ./tests/pw/all-reports/html-report
-            #       destination_dir: ${{ github.event.pull_request.number || github.sha }}
-            #       keep_files: true
-            #       user_name: 'github-actions[bot]' # This will be the username that gets stamped in the repo commit
-            #       user_email: 'github-actions[bot]@users.noreply.github.com' # This will be the user email that gets stamped in the repo com
-
-
-            # - name: Test results
-            #   if: always()
-            #   run: |
-            #       echo "https://$GITHUB_REPOSITORY_OWNER.github.io/${GITHUB_REPOSITORY#*/}"
